+++ src/app/(normal)/post/[source]/[id]/layout.tsx
@@ -2,10 +2,10 @@ import type { Metadata } from 'next';
 import type React from 'react';
 
 import { KeyType } from '@/constants/enum.js';
+import { type SourceInURL } from '@/constants/enum.js';
 import { createSiteMetadata } from '@/helpers/createSiteMetadata.js';
 import { isBotRequest } from '@/helpers/isBotRequest.js';
 import { memoizeWithRedis } from '@/helpers/memoizeWithRedis.js';
-import { type SourceInURL } from '@/helpers/resolveSource.js';
 import { getPostOGById } from '@/services/getPostOGById.js';
 
 const getPostOGByIdRedis = memoizeWithRedis(getPostOGById, {
