+++ src/app/(normal)/post/[source]/[id]/page.tsx
@@ -9,11 +9,11 @@ import { useDocumentTitle } from 'usehooks-ts';
 import ComeBack from '@/assets/comeback.svg';
 import { CommentList } from '@/components/Comments/index.js';
 import { SinglePost } from '@/components/Posts/SinglePost.js';
-import { PageRoutes, SocialPlatform } from '@/constants/enum.js';
+import { PageRoutes, SocialPlatform, SourceInURL } from '@/constants/enum.js';
 import { SITE_NAME } from '@/constants/index.js';
 import { dynamic } from '@/esm/dynamic.js';
 import { createPageTitle } from '@/helpers/createPageTitle.js';
-import { resolveSource, type SourceInURL } from '@/helpers/resolveSource.js';
+import { resolveSource } from '@/helpers/resolveSource.js';
 import { getPostById } from '@/services/getPostById.js';
 import { useGlobalState } from '@/store/useGlobalStore.js';
 import { useImpressionsStore } from '@/store/useImpressionsStore.js';
