+++ src/app/(normal)/profile/[source]/[id]/layout.tsx
@@ -1,11 +1,10 @@
 import type { Metadata } from 'next';
 import type React from 'react';
 
-import { KeyType } from '@/constants/enum.js';
+import { KeyType, type SourceInURL } from '@/constants/enum.js';
 import { createSiteMetadata } from '@/helpers/createSiteMetadata.js';
 import { isBotRequest } from '@/helpers/isBotRequest.js';
 import { memoizeWithRedis } from '@/helpers/memoizeWithRedis.js';
-import { type SourceInURL } from '@/helpers/resolveSource.js';
 import { getProfileOGById } from '@/services/getProfileOGById.js';
 
 const getProfileOGByIdRedis = memoizeWithRedis(getProfileOGById, {
