+++ src/app/api/og/post/route.tsx
@@ -1,6 +1,13 @@
 import { ImageResponse } from 'next/og.js';
+import { z } from 'zod';
 
+import { SourceInURL } from '@/constants/enum.js';
 import { getSatoriFonts } from '@/helpers/getSatoriFonts.js';
+
+const PayloadSchema = z.object({
+    postId: z.string(),
+    source: z.nativeEnum(SourceInURL),
+});
 
 export async function GET(request: Request) {
     return new ImageResponse(<div>Post</div>, {
