+++ src/components/Oembed/Post.tsx
@@ -3,8 +3,8 @@ import { useSuspenseQuery } from '@tanstack/react-query';
 import { memo } from 'react';
 
 import { Quote } from '@/components/Posts/Quote.js';
-import { SocialPlatform } from '@/constants/enum.js';
-import { resolveSource, type SourceInURL } from '@/helpers/resolveSource.js';
+import { SocialPlatform, type SourceInURL } from '@/constants/enum.js';
+import { resolveSource } from '@/helpers/resolveSource.js';
 import { FarcasterSocialMediaProvider } from '@/providers/farcaster/SocialMedia.js';
 import { LensSocialMediaProvider } from '@/providers/lens/SocialMedia.js';
 import { useImpressionsStore } from '@/store/useImpressionsStore.js';
