+++ src/helpers/resolveSource.ts
@@ -1,13 +1,11 @@
 import { createLookupTableResolver } from '@masknet/shared-base';
 
-import { SocialPlatform } from '@/constants/enum.js';
-
-export type SourceInURL = 'lens' | 'farcaster';
+import { SocialPlatform, SourceInURL } from '@/constants/enum.js';
 
 export const resolveSource = createLookupTableResolver<SourceInURL, SocialPlatform>(
     {
-        lens: SocialPlatform.Lens,
-        farcaster: SocialPlatform.Farcaster,
+        [SourceInURL.Farcaster]: SocialPlatform.Farcaster,
+        [SourceInURL.Lens]: SocialPlatform.Lens,
     },
     (keyword) => {
         throw new Error(`Unknown keyword: ${keyword}`);
