+++ src/services/getProfileOGById.ts
@@ -1,10 +1,11 @@
 import urlcat from 'urlcat';
 
+import type { SourceInURL } from '@/constants/enum.js';
 import { SITE_URL } from '@/constants/index.js';
 import { createPageTitle } from '@/helpers/createPageTitle.js';
 import { createSiteMetadata } from '@/helpers/createSiteMetadata.js';
 import { getProfileUrl } from '@/helpers/getProfileUrl.js';
-import { resolveSource, type SourceInURL } from '@/helpers/resolveSource.js';
+import { resolveSource } from '@/helpers/resolveSource.js';
 import { getProfileById } from '@/services/getProfileById.js';
 
 export async function getProfileOGById(source: SourceInURL, profileId: string) {
